first tries with the sbb model. limit candidate set to a certain destination zone
